% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs_filter.R
\docType{data}
\name{featureSelection.filter}
\alias{featureSelection.filter}
\title{Feature selection filter}
\format{An object of class \code{Filter} of length 6.}
\usage{
featureSelection.filter
}
\description{
Creates a filter to select the optimal number of features to fit a riskScorer model.
}
\author{
Mara Tews
}
\keyword{datasets}
