% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_balance.R
\name{multicrit.regularizedRiskScorer}
\alias{multicrit.regularizedRiskScorer}
\title{Multi-criteria tuning for regularizedRiskScorer}
\usage{
multicrit.regularizedRiskScorer(a, min.balance, max.balance, k = 10,
  tune.strategy = mlr::makeTuneMultiCritControlGrid(resolution = 10L),
  logarithmical = FALSE, data, feature.names, y.name, formula,
  importance, weight, beta, ...)
}
\arguments{
\item{a}{[\code{integer}]\cr
Number of SNPs used for inital penalty step, e.g. size of genotyping chip.}

\item{min.balance}{[\code{numeric}]\cr
Lower limit for balance.}

\item{max.balance}{[\code{numeric}]\cr
Upper limit for balance.}

\item{k}{[\code{numeric}]\cr
Number of cross-validation steps. Default is 10.}

\item{tune.strategy}{[\code{TuneMultiCritControl}]\cr
Pass mlr \code{Tune Controll} object to spezify search strategy and resolution.
Search strategy used in cross-validation to optimize \code{balance}. One of those implemented in
\code{\link{mlr}} package. Default is grid search with \code{10L} resolution.}

\item{logarithmical}{[\code{logical}]\cr 
A logical indicating weather penalty step size should be equal (\code{FALSE}) or logarithmic (\code{TRUE}).
Default is \code{TRUE}.}

\item{data}{[\code{\link{data.frame}}]\cr
A data frame (or object coercible by \code{\link{as.data.frame}})
to a data frame) containing the variables in the model.}

\item{feature.names}{[\code{\link{character}}]\cr
A character vector giving the feature names.}

\item{y.name}{[\code{string}]\cr
A string giving the response variable name.}

\item{formula}{[\code{\link{formula}}]\cr
An object of class \code{\link{formula}} (or one that can be
coerced to that class): a symbolic description of the model to
be fitted.}

\item{importance}{[\code{numeric}]\cr
A vector of feature weights to be used in the fitting process.
Should be a named numeric vector.}

\item{weight}{[\code{logical}]\cr
A logical indicating weather weights should be applied
(\code{TRUE}) or used to identify alleles with flipped risk
allele (\code{FALSE}).}

\item{beta}{[\code{logical} or \code{function} or \code{number}]\cr
Depends on \code{weight}: \code{TRUE}, if weights are beta 
coefficients; a \code{function} for transformation to beta
coefficients, if \code{weight} is \code{TRUE}; a \code{number}
giving a threshold as flip criteria, if \code{weight} is
\code{FALSE}.}

\item{...}{Other parameter passed from and to other methods.}
}
\description{
Multi-criteria tuning for balance and number of used features for regularizedRiskScorer model.
}
