% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riskScorer.R
\name{predict.riskScorer}
\alias{predict.riskScorer}
\title{Predict Method for riskScoreR Fits}
\usage{
\method{predict}{riskScorer}(risk.scorer, newdata, type = "score", ...)
}
\arguments{
\item{risk.scorer}{[\code{\link{riskScorer}}]\cr
A fitted object of class "\code{riskScorer}".}

\item{newdata}{[\code{data.frame}]\cr
A data frame in which to look for variables with which to
predict.}

\item{type}{[\code{string}]\cr
The type of prediction required. The default is the
aggregated risk score, i.e. the (un)weighted sum of risk
alleles. The alternative "\code{response}" gives the
predicted binary class labels. The alternative
"\code{prob}" gives the  predicted probabilities of two
classes.}

\item{...}{Other parameter passed from and to other methods.}
}
\value{
Depends on type: if \code{type = "score"} a \code{data.frame} with
one column holding the raw scores, i.e. (un)weighted sum of risk alleles; if
\code{type = "prob"} a \code{data.frame} with two columns, named as the two
factor levels in the response variable of the \code{riskScorer} object, giving
the probabilities of the respective factor level; if \code{type = "response"}
a \code{data.frame} holding the factor level with highest probability.
}
\description{
Obtains predictions from a fitted risk score model object.
}
\details{
Predicted probabilities are obtained by rescaling the distribution
of scores to \eqn{[0,1]}. Predicted class labels are obtained by rounding the
predicted probabilites.
}
