% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizedRiskScorer.R
\name{regularizedRiskScorer}
\alias{regularizedRiskScorer}
\title{Optimization of number of used SNPs.}
\usage{
regularizedRiskScorer(a, balance, data, importance, k, grid.res, weight,
  formula, beta, ...)
}
\arguments{
\item{a}{[\code{integer}]\cr
Initial number of SNPs.}

\item{balance}{[\code{double}]\cr
Weight of penalty.}

\item{data}{[\code{\link{data.frame}}]\cr
A data frame (or object coercible by \code{\link{as.data.frame}})
to a data frame) containing the variables in the model.}

\item{importance}{[\code{numeric}]\cr
A vector of feature weights to be used in the fitting process.
Should be a named numeric vector.}

\item{k}{[\code{integer}]\cr
Number of folds used in cross-validation to find best w. Default is 10.}

\item{grid.res}{[\code{integer}]\cr
Resolution of the grid for w, e.g. number of equal-sized steps. Default is 10.}

\item{...}{Other parameter passed from and to other methods.}
}
\value{
Optimized number of SNPs w to use in final model.
}
\description{
Optimizes number of used SNPs w used in final model based on penaly function.
}
