% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/riskScorer.R
\name{predict.riskScorer}
\alias{predict.riskScorer}
\title{Predict Method for riskScoreR Fits}
\usage{
\method{predict}{riskScorer}(riskScorer, newdata, type = "score", ...)
}
\arguments{
\item{riskScorer}{[\code{\link{riskScorer}}]\cr
A fitted object of class "\code{riskScorer}".}

\item{newdata}{[\code{data.frame}]\cr
A data frame in which to look for variables with which to
predict.}

\item{type}{[\code{string}]\cr
The type of prediction required. The default is the
aggregated risk score, i.e. the (un)weighted sum of risk
alleles. The alternative "\code{response}" gives the
predicted binary class labels. The alternative
"\code{prob}" gives the  predicted probabilities of two
classes.}

\item{...}{Other parameter passed from and to other methods.}
}
\value{
An object of class "\code{data.frame}" is returned with two columns.
First column is the id. If \code{type = "score"} (un)weighted risk scores; if
\code{type = "respons"} predicted class labels; if \code{type = "prob"}
predicted class probabilities are returend in the second column.
}
\description{
Obtains predictions from a fitted risk score model object.
}
\details{
Predicted probabilities are obtained by rescaling the distribution
of scores to \eqn{[0,1]}. Predicted class labels are obtained by rounding the
predicted probabilites.
}

