% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regularizedRiskScorer.R
\name{regularizedRiskScorer}
\alias{regularizedRiskScorer}
\title{Fitting regularized Risk Score Models based on importance of SNPs}
\usage{
regularizedRiskScorer(a, balance, logarithmical, data, importance, weight,
  formula, beta, ...)
}
\arguments{
\item{a}{[\code{integer}]\cr
Number of SNPs used for inital penalty step, e.g. size of genotyping chip.}

\item{balance}{[\code{double}]\cr
Weight of penalty.}

\item{data}{[\code{\link{data.frame}}]\cr
A data frame (or object coercible by \code{\link{as.data.frame}})
to a data frame) containing the variables in the model.}

\item{importance}{[\code{numeric}]\cr
A vector of feature weights to be used in the fitting process.
Should be a named numeric vector.}

\item{weight}{[\code{logical}]\cr
A logical indicating weather weights should be applied
(\code{TRUE}) or used to identify alleles with flipped risk
allele (\code{FALSE}).}

\item{beta}{[\code{logical} or \code{function} or \code{number}]\cr
Depends on \code{weight}: \code{TRUE}, if weights are beta
coefficients; a \code{function} for transformation to beta
coefficients, if \code{weight} is \code{TRUE}; a \code{number}
giving a threshold as flip criteria, if \code{weight} is
\code{FALSE}. See 'Details'.}

\item{...}{Other parameter passed from and to other methods.}

\item{[\code{logarithmical}]\cr}{A logical indicating weather penalty step size should be equal (\code{FALSE}) or logarithmic (\code{TRUE}).
Default is \code{TRUE}.}

\item{y.name}{[\code{string}]\cr
A string giving the response variable name.}

\item{feature.names}{[\code{\link{character}}]\cr
A character vector giving the feature names.}
}
\value{
\code{regularizedRiskScorer} returns an object of class "\code{riskScorer}". For more information see 
\code{\link{riskScorer}}.
}
\description{
Fitting regularized Risk Score Models based on importance of SNPs
}
\details{
Pentalty function can have equal or logarithmic step sizes. Set \code{a} as size of genotyping chip and use 
equal step size to determine the best amount of identical genotyping chips. Set \code{a} as a an inital step size an use 
logarithmic step size to regularize amount of SNPs (preferred all on one chip).
}
